Feat: [EVER-111] 구독 상품 가입 API 만들기

+++ src/main/java/com/team4ever/backend/domain/subscriptions/repository/SubscriptionCombinationRepository.java
@@ -8,6 +8,5 @@ import java.util.Optional;
 
 @Repository
 public interface SubscriptionCombinationRepository extends JpaRepository<SubscriptionCombination, Integer> {
-	Optional<SubscriptionCombination> findBySubscriptionIdAndBrandIdAndUserId(
-			Integer subscriptionId, Integer brandId, Integer userId);
+	Optional<SubscriptionCombination> findBySubscriptionIdAndBrandIdAndUserId(Integer subscriptionId, Integer brandId, Long userId);
 }
